selesai versi 1

+++ app/Http/Controllers/AspectController.php
@@ -9,21 +9,24 @@ use App\Models\Master;
 
 class AspectController extends Controller
 {
-    public function home()
+    public function home($tahun)
     {
-        $master = Master::where('jenis','Aspek')->get();
-        return view('master.aspek',['master'=>$master]);
+        $master = Master::where('jenis','Aspek')->where('tahun',$tahun)->get();
+        // return $tahun;
+        return view('master.aspek',['master'=>$master,'tahun'=>$tahun]);
     }
-    public function add(Request $req)
+    public function add(Request $req,$tahun)
     {
-        $count = count(Master::where('jenis','Aspek')->get())+1;
+        $count = count(Master::where('jenis','Aspek')->where('tahun',$tahun)->get())+1;
         
         $master = new Master();
         $master->urutan = $count;
         $master->nama = $req->nama;
         $master->jenis = "Aspek";
         $master->bobot = $req->bobot;
+        $master->skor = 0;
         $master->catatan = $req->catatan;
+        $master->tahun = $tahun;
         $master->save();
         return redirect()->back()->with(['success' => 'Pesan Berhasil']);
     }
@@ -40,12 +43,18 @@ class AspectController extends Controller
         $master->bobot = $req->bobot;
         $master->catatan = $req->catatan;
         $master->save();
-        return redirect('/aspek')->with(['success' => 'Pesan Berhasil']);
+        $tahun = $master->tahun;
+        return redirect('/aspek/'.$tahun)->with(['success' => 'Pesan Berhasil']);
     }
     public function delete($id)
     {
         $master = Master::find($id);
-        $master->delete();
-        return redirect('/aspek')->with(['warning' => 'Pesan Berhasil']);
+        $child = Master::where('jenis','Indikator')->where('id_parent',$master->id)->get();
+        if (count($child) == 0) {
+            $master->delete();
+            return redirect()->back()->with(['warning' => 'Pesan Berhasil']);
+        } else {
+            return redirect()->back()->with(['info' => 'Gagal Hapus']);
+        }
     }
 }
